Changes many things.... cant describe

+++ php/classes/experince.php
@@ -20,21 +20,38 @@ $this->description=$description;
 function insertIntoDatabase()
 {
 $db = (new Database())->connectToDatabase();
-//$db->insert('user_experince',array($this->username,$this->title,$this->description));
+$db->query("SELECT * FROM user_experince where experinceid='$this->experinceid'");
+$rows=$db->returned_rows;
+if($rows==1)
+{
+	return array('status_code'=>400 , 'detail'=> "already exists");
+}
+else 
+{
 $db->query("INSERT INTO user_experince values('$this->username','this->experinceid','$this->title','$this->description')");
+return array('status_code'=>200);
+}
 }
 
-
-function deleteDescription()
+function deleteExperince()
 {
 $db = (new Database())->connectToDatabase();
+$db->query("SELECT * FROM user_experince where experinceid='$this->experinceid'");
+$rows=$db->returned_rows;
+echo "<br>".$rows;
+if($rows==1)
+{
 $db->query("DELETE from user_experince where experinceid=$this->experinceid");
+return array('status_code'=>200);
+}
+else 
+return array('status_code'=>404, 'detail'=>'comment not found');
 }
 
 
 }
 
-$eid=new experince(1,123,1234,1);
-$eid->insertIntoDatabase();
-
+$eid=new experince(11,123,12354,7);
+$eid->deleteExperince();
+echo $eid->deleteExperince()['detail'];
 ?>
